added unit tests for reload
- Auto_reload_validxml_test succeeds
- Auto_Reload_invalidxml_test fails
